+++ collector/src/autodoc_collector/k8s_apis.py
@@ -19,6 +19,8 @@ class K8sApis:
     custom_objects: client.CustomObjectsApi
     autoscaling_v2: client.AutoscalingV2Api
     storage_v1: client.StorageV1Api
+    rbac_v1: client.RbacAuthorizationV1Api
+    policy_v1: client.PolicyV1Api
 
     @classmethod
     def build(cls) -> K8sApis:
@@ -30,4 +32,6 @@ class K8sApis:
             custom_objects=client.CustomObjectsApi(),
             autoscaling_v2=client.AutoscalingV2Api(),
             storage_v1=client.StorageV1Api(),
+            rbac_v1=client.RbacAuthorizationV1Api(),
+            policy_v1=client.PolicyV1Api(),
         )